Refactoring:
Eliminate dal modello delle entit?, tutte le classi per gestire le Correzioni.
Sono tornato alla ipotesi di fare una normale collezioni di oggetti e serializzarla come stringa xml
in un campo del database diretto.

Eliminato il Provinatore in favore di una Correzione di tipo Resize.

Eliminate le classi "Comando" e derivate. Ora faccio tutto dentro il servizio apposito IFotoRitoccoSrv.

Creata nuova schermata per il fotoritocco e implementato il Bianco-Nero, Ruota, Torna-Originale
  (piccolo bug: le correzioni non vengono correttamente scritte nel db)

+++ Digiphoto.Lumen.Model/src/Digiphoto/Lumen/Imaging/IImmagine.cs
@@ -11,11 +11,11 @@ namespace Digiphoto.Lumen.Imaging {
 
 	public interface IImmagine : IDisposable {
 
-		int ww { 
+		long ww { 
 			get; 
 		}
 		
-		int hh { 
+		long hh { 
 			get; 
 		}
